pyproject.toml
[project]
name = "roguelike"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "tcod>=16.0.0",
    "pyyaml>=6.0.0",
]
